requirements.txt
discord.py==2.2.2
jishaku==2.5.1
python-dotenv==1.0.0